+++ Network/API/Telegram/Bot/Object.hs
@@ -3,3 +3,4 @@ module Network.API.Telegram.Bot.Object (module Exports) where
 import Network.API.Telegram.Bot.Object.Update as Exports
 import Network.API.Telegram.Bot.Object.Sender as Exports
 import Network.API.Telegram.Bot.Object.Member as Exports
+import Network.API.Telegram.Bot.Object.Chat as Exports