+++ Network/API/Telegram/Bot/Object/Chat.hs
@@ -1,0 +1,5 @@
+module Network.API.Telegram.Bot.Object.Chat (module Exports) where
+
+import Network.API.Telegram.Bot.Object.Chat.Channel as Exports
+import Network.API.Telegram.Bot.Object.Chat.Conversation as Exports
+import Network.API.Telegram.Bot.Object.Chat.Group as Exports